Added a method for getting all matches from a league other than the ones you already have.
Extended the wait time between failed calls.

Several minor bug fixes.

+++ Data/Player.cs
@@ -5,7 +5,7 @@ namespace SteamWebAPIWrapper.Data
     public class Player
     {
         [JsonProperty(PropertyName = "account_id")]
-        public int AccountId { get; set; }
+        public long AccountId { get; set; }
 
         [JsonProperty(PropertyName = "player_slot")]
         public uint PlayerSlot { get; set; }
